Rename bar to nodeBar as its a specific bar implementation
